requirements.txt
langchain
phi==0.6.7
pgvector==0.3.6
psycopg[binary]
pypdf
sqlalchemy
openai
packaging
duckduckgo-search
phidata==2.7.10
rich==13.9.4
pydantic==2.10.6
python-dotenv
gitpython==3.1.44
uvicorn
httpx==0.28.1
fastapi
groq





